arch:xtens:mpu: modify acc and memtype to uint32_t
The uint8_t and uint16_t will overflow in MPU_ENTRY_AR marco.

+++ arch/xtensa/src/common/mpu.h
@@ -210,7 +210,7 @@ void mpu_control(bool enable);
  ****************************************************************************/
 
 void mpu_configure_region(uintptr_t base, size_t size,
-            uint8_t acc, uint16_t memtype);
+                          uint32_t acc, uint32_t memtype);
 
 /****************************************************************************
  * Name: mpu_priv_stronglyordered
